Parsing functions moved to parsing.h
