+++ src/rws_client.cpp
@@ -36,6 +36,7 @@
 
 #include <abb_librws/rws_client.h>
 #include <abb_librws/rws_error.h>
+#include <abb_librws/parsing.h>
 
 #include <Poco/Net/HTTPRequest.h>
